extract masked_multi_head_self_attention from utils to module
